Move AdvancedSettingsFragment to a new file
