+++ AnkiDroid/src/main/res/xml/preference_headers.xml
@@ -79,7 +79,7 @@
 
     <!-- Advanced Preferences -->
     <Preference
-        android:fragment="com.ichi2.anki.Preferences$AdvancedSettingsFragment"
+        android:fragment="com.ichi2.anki.preferences.AdvancedSettingsFragment"
         android:key="@string/pref_advanced_screen_key"
         android:summary="@string/pref_cat_advanced_summ"
         android:title="@string/pref_cat_advanced"
